mmc: refactor card recovery improvements
This commit refactors the "reset and retry or give up" algorithm
from K3.0 to better align it with the existing "reset and retry"
functionality from upstream.  This approach makes use of the
existing framework instead of attempting to replace it.

The logic used to determine when to give up on a card
has been improved.  Instead of giving up after a set number of
failures, we now track the failure rate when trouble starts and
forgive cards that behave themselves for a period of time.  This
improves the user experience for extremely low-cost cards that
tend to misbehave in small spurts and then work perfectly fine
for hours or days.

Change-Id: I00976e4d4a3523dfff4a18d42b1f0826f1fda6e3
Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/613547
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

Conflicts:
	drivers/mmc/card/block.c