mmc: add sysfs knobs for card recovery logic
This commit exports the failure and forgivness ratios to sysfs, so
that the values can be tweaked from user space.  Problematic hosts
may wish to relax the parameters under which we drop an error-
producing card, for example.  The defaults ratios are 10 for both.

Change-Id: I20b4048b6870f6b62b710fb3e67b8937447cd6e6
Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/619528
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Igor Kovalenko <igork@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>